:sparkles: add login and logut method to driver controller, update routes file

+++ db/migrate/20240526080129_create_drivers.rb
@@ -3,9 +3,10 @@ class CreateDrivers < ActiveRecord::Migration[7.0]
     create_table :drivers do |t|
       t.string :email
       t.string :password_digest
-      t.integer :phone
+      t.bigint :phone
       t.integer :vehicle_id
       t.integer :location_id
+      
 
       t.timestamps
     end
